backend/requirements.txt
chromadb~=1.3.4
#FASTAPI libraries for api enablement
fastapi~=0.120.1
uvicorn~=0.38.0

# Langchain libraries for various functionalities
langchain~=1.0.7
langchain-community~=0.4.1
langchain-chroma~=1.0.0  # Chroma integration for Langchain
langchain_core~=1.0.5 # Core functionalities of Langchain
langchain-ollama~=1.0.0  # Ollama integration for Langchain
langchain-text-splitters~=1.0.0  # Text splitting utilities for Langchain

#agent orchestration
langgraph~=1.0.3  # Graph-based agent management

# Natural Language Processing libraries
dspy~=3.0.3
nltk~=3.9.2  # Natural Language Toolkit for text processing
ollama~=0.6.0  # Ollama library for LLM operations
pandas~=2.3.3  # Data manipulation and analysis library

# Data validation and environment management
praw~=7.8.1 # Python Reddit API Wrapper
pydantic~=2.12.3  # Data validation and settings management
python-certifi-win32~=1.6.1  # SSL certificate validation for Windows
python-dotenv~=1.2.1  # Load environment variables from .env files

# Machine learning and data science libraries
scikit-learn~=1.7.2  # Machine learning library
sentence-transformers~=5.1.2  # Sentence embeddings for NLP tasks

# Documentation generation tools
sphinx~=8.2.3  # Documentation generator
sphinx-autodoc-typehints~=3.5.2  # Autodoc extension for type hints

# Text summarization and sentiment analysis
sumy~=0.11.0  # Text summarization library
textblob~=0.19.0  # Simple NLP library for processing textual data
vaderSentiment~=3.3.2  # Sentiment analysis tool for social media text

# Multi-Provider LLM Support (for generic_llm_interface.py)
# Install only the providers you plan to use

# OpenAI support
langchain-openai~=1.0.3  # ChatOpenAI, OpenAIEmbeddings

# Anthropic (Claude) support
langchain-anthropic>=1.1.0  # ChatAnthropic

# Google (Gemini) support
langchain-google-genai>=3.1.0  # ChatGoogleGenerativeAI, GoogleGenerativeAIEmbeddings

# HuggingFace support
langchain-huggingface>=0.1.0  # HuggingFaceEndpoint, HuggingFaceEmbeddings
